+++ lib/src/core/logger.dart
@@ -2,12 +2,12 @@ import 'dart:developer' show log;
 import 'package:flutter_super/src/injection.dart';
 
 /// The default Logger of the Super Framework
-void logger(String msg, {bool warning = false}) {
+void logger(Object object, {bool warning = false}) {
   if (Injection.enableLog) {
     if (warning) {
-      log(msg, name: 'Super');
+      log('$object', name: 'Super');
     } else {
-      log('\x1B[36m$msg\x1B[0m', name: 'Super');
+      log('\x1B[36m$object\x1B[0m', name: 'Super');
     }
   }
 }