+++ test/src/rx/rx_test.dart
@@ -8,7 +8,7 @@ void main() {
       final rx1 = RxInt(10);
       final rx2 = RxString('Hello');
 
-      final mergedRx = RxMerge([rx1, rx2]);
+      final mergedRx = RxMerge<Object>([rx1, rx2]);
 
       expect('${mergedRx.children}', '[$rx1, $rx2]');
 
@@ -26,7 +26,7 @@ void main() {
       });
 
       expect('${mergedRx.children}', '[$rx1, $rx2]');
-      expect(mergedRx.toString(), '${RxMerge([rx1, rx2])}');
+      expect(mergedRx.toString(), '${RxMerge<Object>([rx1, rx2])}');
       expect(listenerCalled, true);
     });
   });